refactor(http-client): reorganize tests to 1:1 file structure
- Removed __tests__ directory
- Created individual test files for each contract:
  - client.test-d.ts
  - response.test-d.ts
  - errors.test-d.ts
  - interceptors.test-d.ts
  - request.test-d.ts
  - retry.test-d.ts
  - factory.test-d.ts
- Moved imports before JSDoc comments in all test files
- Fixed ESLint errors with inline eslint-disable comments
- All 66 type tests + 6 unit tests passing
